Display requested ad units
Add method to display the ad units that have been included on another HTML page via a script tag. The content-type header is set to Javascript and we use document.write to create the ad unit.

The world just doesn't have enough document.write in it yet, so we're contributing.